ingestion/requirements.txt
dlt[bigquery]==0.4.12
eurostat>=1.0.0
pandas==2.2.2
numpy==1.26.4
kafka-python==2.0.2
google-cloud-bigquery==3.17.2
google-cloud-storage==2.16.0
pyarrow==15.0.2
requests==2.32.3
python-dotenv==1.0.1
